+++ docs/cda-logical-model/assets/js/anchor-hover.js
@@ -1,0 +1,2 @@
+anchors.options.visible = 'hover'
+anchors.add()