+++ docs/cda-logical-model/assets/js/window-hash.js
@@ -1,0 +1,5 @@
+$(document).ready(function(){
+  if(window.location.hash != "") {
+      $('a[href="' + window.location.hash + '"]').click()
+  }
+});