+++ matchbox-engine/src/main/java/org/hl7/fhir/utilities/npm/FilesystemPackageCacheManager.java
@@ -518,7 +518,7 @@ public class FilesystemPackageCacheManager extends BasePackageCacheManager imple
   	}
 
     if ("hl7.cda.uv.core".equals(id)) {
-  		version = "2.0.0-sd-202312-matchbox-patch";
+  		version = "2.0.0-sd-202406-matchbox-patch";
       NpmPackage npm = NpmPackage.fromPackage(packageTgzInputStream, sourceDesc, true);
       return npm;
   	}